// packages/mcp-cafe24-admin/src/types/cancellation.ts
export interface Cancellation {
  shop_no: number;
  order_id: string;
  claim_code: string;
  claim_reason_type: string;
  claim_reason: string;
  refund_methods: string[];
  refund_reason: string;
  order_price_amount: string;
  refund_amounts: {
    payment_method: string;
    amount: string;
  }[];
  shipping_fee: string;
  return_ship_type: string;
  defer_commission: string;
  partner_discount_amount: string;
  add_discount_amount: string;
  member_grade_discount_amount: string;
  shipping_discount_amount: string;
  coupon_discount_amount: string;
  point_used: string;
  credit_used: string;
  undone: string;
  undone_reason_type: string | null;
  undone_reason: string | null;
  expose_order_detail: string | null;
  exposed_undone_reason: string | null;
  items: CancellationItem[];
  include_tax: string;
  tax:
    | {
        name: string;
        amount: string;
      }[]
    | null;
  cancel_fee_amount: string | null;
}

export interface CancellationItem {
  shop_no: number;
  item_no: number;
  order_item_code: string;
  variant_code: string;
  product_no: number;
  product_code: string;
  custom_product_code: string;
  custom_variant_code: string;
  eng_product_name: string | null;
  option_id: string;
  option_value: string;
  option_value_default: string;
  additional_option_value: string;
  additional_option_values: {
    key: string;
    type: string;
    name: string;
    value: string;
  }[];
  product_name: string;
  product_name_default: string;
  product_price: string;
  option_price: string;
  additional_discount_price: string;
  coupon_discount_price: string;
  app_item_discount_amount: string;
  quantity: number;
  supplier_product_name: string;
  supplier_transaction_type: string;
  supplier_id: string;
  supplier_name: string;
  tracking_no: string;
  shipping_code: string;
  claim_code: string;
  claim_reason_type: string;
  claim_reason: string;
  refund_bank_name: string;
  refund_bank_account_no: string;
  refund_bank_account_holder: string;
  post_express_flag: string | null;
  order_status: string;
  order_status_additional_info: string | null;
  claim_quantity: number;
  status_code: string;
  status_text: string;
  open_market_status: string;
  bundled_shipping_type: string;
  shipping_company_id: string;
  shipping_company_name: string;
  shipping_company_code: string;
  product_bundle: string;
  product_bundle_no: string;
  product_bundle_name: string | null;
  product_bundle_name_default: string | null;
  product_bundle_type: string;
  was_product_bundle: string | null;
  original_bundle_item_no: number | null;
  naver_pay_order_id: string | null;
  naver_pay_claim_status: string;
  individual_shipping_fee: string;
  shipping_fee_type: string;
  shipping_fee_type_text: string;
  shipping_payment_option: string;
  payment_info_id: string;
  original_item_no: number[];
  store_pickup: string;
  ordered_date: string;
  shipped_date: string;
  delivered_date: string;
  cancel_date: string | null;
  return_request_date: string | null;
  return_confirmed_date: string | null;
  return_collected_date: string | null;
  cancel_request_date: string | null;
  refund_date: string | null;
  exchange_request_date: string | null;
  exchange_date: string | null;
  product_material: string | null;
  product_material_eng: string | null;
  cloth_fabric: string | null;
  product_weight: string;
  volume_size: string | null;
  volume_size_weight: string | null;
  clearance_category: string | null;
  clearance_category_info: string | null;
  clearance_category_code: string | null;
  hs_code: string;
  one_plus_n_event: string | null;
  origin_place: string;
  gift: string;
  item_granting_gift: string | null;
  product_bundle_list: CancellationBundleItem[];
  market_cancel_request: string | null;
  market_cancel_request_quantity: number | null;
  market_fail_reason: string | null;
  market_fail_reason_guide: string | null;
  market_item_sequence: number | null;
  market_item_no: string | null;
  market_custom_variant_code: string | null;
  option_type: string;
  options: {
    option_code: string;
    option_name: string | null;
    option_value: {
      option_text: string | null;
      value_no: number;
    };
  }[];
  market_discount_amount: string;
  labels: string | null;
  order_status_before_cs: string;
  supply_price: string;
  multi_invoice: string | null;
  shipping_expected_date: string | null;
}

export interface CancellationBundleItem {
  product_no: number;
  product_code: string;
  variant_code: string;
  product_name: string;
  product_name_default: string;
  option_id: string;
  option_value: string;
  option_value_default: string;
  additional_option_value: string;
  additional_option_values: {
    key: string;
    type: string;
    name: string;
    value: string;
  }[];
  quantity: number;
  supplier_id: string;
  eng_product_name: string | null;
  hs_code: string;
  option_price: string;
}